+++ src/main/groovy/de/laser/api/v0/entities/ApiCostItem.groovy
@@ -21,7 +21,7 @@ class ApiCostItem {
 
     /**
      * Locates the given {@link CostItem} and returns the object (or null if not found) and the request status for further processing
-     * @param the field to look for the identifier, one of {id, globalUID}
+     * @param the field to look for the identifier, one of {id, laserID}
      * @param the identifier value
      * @return {@link ApiBox}(obj: CostItem | null, status: null | BAD_REQUEST | PRECONDITION_FAILED | NOT_FOUND | OBJECT_STATUS_DELETED)
      * @see ApiBox#validatePrecondition_1()
@@ -38,7 +38,7 @@ class ApiCostItem {
                     }
                 }
                 break
-            case 'globalUID':
+            case 'laserID':
                 result.obj = CostItem.findAllByGlobalUID(value)
                 if(!result.obj) {
                     DeletedObject.withTransaction {
@@ -160,7 +160,7 @@ class ApiCostItem {
 
         costItem = GrailsHibernateUtil.unwrapIfProxy(costItem)
 
-        result.globalUID           = costItem.globalUID
+        result.laserID             = costItem.globalUID
 
         result.costInBillingCurrency            = costItem.costInBillingCurrency
         result.costInBillingCurrencyAfterTax    = costItem.costInBillingCurrencyAfterTax