+++ src/main/groovy/de/laser/api/v0/entities/ApiIssueEntitlement.groovy
@@ -77,7 +77,7 @@ class ApiIssueEntitlement {
             return null
         }
 
-        result.globalUID        = ie.globalUID
+        result.laserID          = ie.globalUID
         //result.name             = ie.name
         result.accessStartDate  = ApiToolkit.formatInternalDate(ie.accessStartDate)
         result.accessEndDate    = ApiToolkit.formatInternalDate(ie.accessEndDate)