+++ src/main/groovy/de/laser/api/v0/entities/ApiLibrarySupplier.groovy
@@ -19,7 +19,7 @@ class ApiLibrarySupplier {
 
     /**
      * Locates the given {@link Vendor} and returns the object (or null if not found) and the request status for further processing
-     * @param the field to look for the identifier, one of {id, globalUID, gokbId, ns:identifier}
+     * @param the field to look for the identifier, one of {id, laserID, gokbId, ns:identifier}
      * @param the identifier value
      * @return {@link ApiBox}(obj: Vendor | null, status: null | BAD_REQUEST | PRECONDITION_FAILED | NOT_FOUND | OBJECT_STATUS_DELETED)
      * @see ApiBox#validatePrecondition_1()
@@ -39,7 +39,7 @@ class ApiLibrarySupplier {
                     }
                 }
                 break
-            case 'globalUID':
+            case 'laserID':
                 result.obj = Vendor.findAllByGlobalUID(value)
                 if(!result.obj) {
                     DeletedObject.withTransaction {