+++ src/main/groovy/de/laser/api/v0/entities/ApiPkg.groovy
@@ -18,7 +18,7 @@ class ApiPkg {
 
     /**
 	 * Locates the given {@link Package} and returns the object (or null if not found) and the request status for further processing
-	 * @param the field to look for the identifier, one of {id, globalUID, gokbId, ns:identifier}
+	 * @param the field to look for the identifier, one of {id, laserID, gokbId, ns:identifier}
 	 * @param the identifier value
      * @return {@link ApiBox}(obj: Package | null, status: null | BAD_REQUEST | PRECONDITION_FAILED | NOT_FOUND | OBJECT_STATUS_DELETED)
 	 * @see ApiBox#validatePrecondition_1()
@@ -31,7 +31,7 @@ class ApiPkg {
 				case 'id':
 					result.obj = Package.get(value)
 					break
-				case 'globalUID':
+				case 'laserID':
 					result.obj = Package.findByGlobalUID(value)
 					break
 				case 'wekbId':
@@ -83,7 +83,7 @@ class ApiPkg {
 
 		pkg = GrailsHibernateUtil.unwrapIfProxy(pkg)
 
-		result.globalUID        	= pkg.globalUID
+		result.laserID          	= pkg.globalUID
 		result.wekbId           	= pkg.gokbId
 		result.name             	= pkg.name
 		result.altnames     		= ApiCollectionReader.getAlternativeNameCollection(pkg.altnames)