+++ src/main/groovy/de/laser/api/v0/entities/ApiProvider.groovy
@@ -15,7 +15,7 @@ class ApiProvider {
 
     /**
      * Locates the given {@link de.laser.wekb.Provider} and returns the object (or null if not found) and the request status for further processing
-     * @param the field to look for the identifier, one of {id, globalUID, gokbId, ns:identifier}
+     * @param the field to look for the identifier, one of {id, laserID, gokbId, ns:identifier}
      * @param the identifier value
      * @return {@link ApiBox}(obj: Provider | null, status: null | BAD_REQUEST | PRECONDITION_FAILED | NOT_FOUND | OBJECT_STATUS_DELETED)
      * @see ApiBox#validatePrecondition_1()
@@ -35,7 +35,7 @@ class ApiProvider {
                     }
                 }
                 break
-            case 'globalUID':
+            case 'laserID':
                 result.obj = Provider.findByGlobalUID(value)
                 if(!result.obj) {
                     DeletedObject.withTransaction {